Split render validation
